Renamed package to dao, handled SQLExceptions inside the methods, created getInstance method to have only one instance of the ConnectionPool, and deleted unused constructors and plugins.
