+++ src/main/java/com/solvd/app/interfaces/IBaseDAO.java
@@ -1,16 +1,15 @@
 package com.solvd.app.interfaces;
 
-import java.sql.SQLException;
 import java.util.List;
 
 public interface IBaseDAO<T> {
-    void createEntity(T t) throws SQLException;
+    void createEntity(T t);
 
-    T getEntityByID(int id) throws SQLException;
+    T getEntityByID(int id);
 
-    void updateEntity(T t) throws SQLException;
+    void updateEntity(T t);
 
-    void deleteEntityByID(int id) throws SQLException;
+    void deleteEntityByID(int id);
 
-    List<T> getAll() throws SQLException;
+    List<T> getAll();
 }
