+++ src/main/java/com/solvd/app/interfaces/IPersonDAO.java
@@ -2,8 +2,6 @@ package com.solvd.app.interfaces;
 
 import com.solvd.app.models.Person;
 
-import java.sql.SQLException;
-
 public interface IPersonDAO extends IBaseDAO<Person> {
-    Person getPersonByNameAndPhoneNumber(String name, int phoneNumber) throws SQLException;
+    Person getPersonByNameAndPhoneNumber(String name, int phoneNumber);
 }
