+++ src/main/java/com/solvd/app/interfaces/IPrescriptionItemDAO.java
@@ -3,9 +3,8 @@ package com.solvd.app.interfaces;
 import com.solvd.app.models.Prescription;
 import com.solvd.app.models.PrescriptionItem;
 
-import java.sql.SQLException;
 import java.util.List;
 
 public interface IPrescriptionItemDAO extends IBaseDAO<PrescriptionItem> {
-    List<PrescriptionItem> getItemsByPrescription(Prescription prescription) throws SQLException;
+    List<PrescriptionItem> getItemsByPrescription(Prescription prescription);
 }
