+++ src/main/java/com/solvd/app/services/DoctorService.java
@@ -1,35 +1,31 @@
 package com.solvd.app.services;
 
-import com.solvd.app.daos.DoctorDAO;
+import com.solvd.app.dao.DoctorDAO;
 import com.solvd.app.models.Doctor;
 
 import java.sql.SQLException;
 import java.util.List;
 
 public class DoctorService {
-    private DoctorDAO doctorDAO;
+    private DoctorDAO doctorDAO = new DoctorDAO();
 
-    public DoctorService(DoctorDAO doctorDAO) {
-        this.doctorDAO = doctorDAO;
-    }
-
-    public void createDoctor(Doctor doctor) throws SQLException {
+    public void createDoctor(Doctor doctor) {
         doctorDAO.createEntity(doctor);
     }
 
-    public Doctor getDoctorByID(int id) throws SQLException {
+    public Doctor getDoctorByID(int id) {
         return doctorDAO.getEntityByID(id);
     }
 
-    public void updateDoctor(Doctor doctor) throws SQLException {
+    public void updateDoctor(Doctor doctor) {
         doctorDAO.updateEntity(doctor);
     }
 
-    public void deleteDoctorByID(int id) throws SQLException {
+    public void deleteDoctorByID(int id) {
         doctorDAO.deleteEntityByID(id);
     }
 
-    public List<Doctor> getAllDoctors() throws SQLException {
+    public List<Doctor> getAllDoctors() {
         return doctorDAO.getAll();
     }
 }
