+++ src/main/java/com/solvd/app/services/DrugCategoryService.java
@@ -1,35 +1,30 @@
 package com.solvd.app.services;
 
-import com.solvd.app.daos.DrugCategoryDAO;
+import com.solvd.app.dao.DrugCategoryDAO;
 import com.solvd.app.models.DrugCategory;
 
-import java.sql.SQLException;
 import java.util.List;
 
 public class DrugCategoryService {
-    private DrugCategoryDAO drugCategoryDAO;
+    private DrugCategoryDAO drugCategoryDAO = new DrugCategoryDAO();
 
-    public DrugCategoryService(DrugCategoryDAO drugCategoryDAO) {
-        this.drugCategoryDAO = drugCategoryDAO;
-    }
-
-    public void createDrugCategory(DrugCategory drugCategory) throws SQLException {
+    public void createDrugCategory(DrugCategory drugCategory) {
         drugCategoryDAO.createEntity(drugCategory);
     }
 
-    public DrugCategory getDrugCategoryByID(int id) throws SQLException {
+    public DrugCategory getDrugCategoryByID(int id) {
         return drugCategoryDAO.getEntityByID(id);
     }
 
-    public void updateDrugCategory(DrugCategory drugCategory) throws SQLException {
+    public void updateDrugCategory(DrugCategory drugCategory) {
         drugCategoryDAO.updateEntity(drugCategory);
     }
 
-    public void deleteDrugCategoryByID(int id) throws SQLException {
+    public void deleteDrugCategoryByID(int id) {
         drugCategoryDAO.deleteEntityByID(id);
     }
 
-    public List<DrugCategory> getAllDrugCategories() throws SQLException {
+    public List<DrugCategory> getAllDrugCategories() {
         return drugCategoryDAO.getAll();
     }
 }
