+++ src/main/java/com/solvd/app/services/InventoryService.java
@@ -1,35 +1,30 @@
 package com.solvd.app.services;
 
-import com.solvd.app.daos.InventoryDAO;
+import com.solvd.app.dao.InventoryDAO;
 import com.solvd.app.models.Inventory;
 
-import java.sql.SQLException;
 import java.util.List;
 
 public class InventoryService {
-    private InventoryDAO inventoryDAO;
+    private InventoryDAO inventoryDAO = new InventoryDAO();
 
-    public InventoryService(InventoryDAO inventoryDAO) {
-        this.inventoryDAO = inventoryDAO;
-    }
-
-    public void createInventory(Inventory inventory) throws SQLException {
+    public void createInventory(Inventory inventory) {
         inventoryDAO.createEntity(inventory);
     }
 
-    public Inventory getInventoryByID(int id) throws SQLException {
+    public Inventory getInventoryByID(int id) {
         return inventoryDAO.getEntityByID(id);
     }
 
-    public void updateInventory(Inventory inventory) throws SQLException {
+    public void updateInventory(Inventory inventory) {
         inventoryDAO.updateEntity(inventory);
     }
 
-    public void deleteInventoryByID(int id) throws SQLException {
+    public void deleteInventoryByID(int id) {
         inventoryDAO.deleteEntityByID(id);
     }
 
-    public List<Inventory> getAllInventories() throws SQLException {
+    public List<Inventory> getAllInventories() {
         return inventoryDAO.getAll();
     }
 }
