+++ src/main/java/com/solvd/app/services/ManufacturerService.java
@@ -1,35 +1,30 @@
 package com.solvd.app.services;
 
-import com.solvd.app.daos.ManufacturerDAO;
+import com.solvd.app.dao.ManufacturerDAO;
 import com.solvd.app.models.Manufacturer;
 
-import java.sql.SQLException;
 import java.util.List;
 
 public class ManufacturerService {
-    private ManufacturerDAO manufacturerDAO;
+    private ManufacturerDAO manufacturerDAO = new ManufacturerDAO();
 
-    public ManufacturerService(ManufacturerDAO manufacturerDAO) {
-        this.manufacturerDAO = manufacturerDAO;
-    }
-
-    public void createManufacturer(Manufacturer manufacturer) throws SQLException {
+    public void createManufacturer(Manufacturer manufacturer) {
         manufacturerDAO.createEntity(manufacturer);
     }
 
-    public Manufacturer getManufacturerByID(int id) throws SQLException {
+    public Manufacturer getManufacturerByID(int id) {
         return manufacturerDAO.getEntityByID(id);
     }
 
-    public void updateManufacturer(Manufacturer manufacturer) throws SQLException {
+    public void updateManufacturer(Manufacturer manufacturer) {
         manufacturerDAO.updateEntity(manufacturer);
     }
 
-    public void deleteManufacturerByID(int id) throws SQLException {
+    public void deleteManufacturerByID(int id) {
         manufacturerDAO.deleteEntityByID(id);
     }
 
-    public List<Manufacturer> getAllManufacturers() throws SQLException {
+    public List<Manufacturer> getAllManufacturers() {
         return manufacturerDAO.getAll();
     }
 }
