+++ src/main/java/com/solvd/app/services/PersonService.java
@@ -1,39 +1,34 @@
 package com.solvd.app.services;
 
-import com.solvd.app.daos.PersonDAO;
+import com.solvd.app.dao.PersonDAO;
 import com.solvd.app.models.Person;
 
-import java.sql.SQLException;
 import java.util.List;
 
 public class PersonService {
-    private PersonDAO personDAO;
+    private PersonDAO personDAO = new PersonDAO();
 
-    public PersonService(PersonDAO personDAO) {
-        this.personDAO = personDAO;
-    }
-
-    public void createPerson(Person person) throws SQLException {
+    public void createPerson(Person person) {
         personDAO.createEntity(person);
     }
 
-    public Person getPersonByID(int id) throws SQLException {
+    public Person getPersonByID(int id) {
         return personDAO.getEntityByID(id);
     }
 
-    public Person getPersonByNameAndPhoneNumber(String name, int phoneNumber) throws SQLException {
+    public Person getPersonByNameAndPhoneNumber(String name, int phoneNumber) {
         return personDAO.getPersonByNameAndPhoneNumber(name, phoneNumber);
     }
 
-    public void updatePerson(Person person) throws SQLException {
+    public void updatePerson(Person person) {
         personDAO.updateEntity(person);
     }
 
-    public void deletePersonByID(int id) throws SQLException {
+    public void deletePersonByID(int id) {
         personDAO.deleteEntityByID(id);
     }
 
-    public List<Person> getAllPersons() throws SQLException {
+    public List<Person> getAllPersons() {
         return personDAO.getAll();
     }
 }
