+++ src/main/java/com/solvd/app/services/PharmacyService.java
@@ -1,35 +1,30 @@
 package com.solvd.app.services;
 
-import com.solvd.app.daos.PharmacyDAO;
+import com.solvd.app.dao.PharmacyDAO;
 import com.solvd.app.models.Pharmacy;
 
-import java.sql.SQLException;
 import java.util.List;
 
 public class PharmacyService {
-    private PharmacyDAO pharmacyDAO;
+    private PharmacyDAO pharmacyDAO = new PharmacyDAO();
 
-    public PharmacyService(PharmacyDAO pharmacyDAO) {
-        this.pharmacyDAO = pharmacyDAO;
-    }
-
-    public void createPharmacy(Pharmacy pharmacy) throws SQLException {
+    public void createPharmacy(Pharmacy pharmacy) {
         pharmacyDAO.createEntity(pharmacy);
     }
 
-    public Pharmacy getPharmacyByID(int id) throws SQLException {
+    public Pharmacy getPharmacyByID(int id) {
         return pharmacyDAO.getEntityByID(id);
     }
 
-    public void updatePharmacy(Pharmacy pharmacy) throws SQLException {
+    public void updatePharmacy(Pharmacy pharmacy) {
         pharmacyDAO.updateEntity(pharmacy);
     }
 
-    public void deletePharmacyByID(int id) throws SQLException {
+    public void deletePharmacyByID(int id) {
         pharmacyDAO.deleteEntityByID(id);
     }
 
-    public List<Pharmacy> getAllPharmacies() throws SQLException {
+    public List<Pharmacy> getAllPharmacies() {
         return pharmacyDAO.getAll();
     }
 }
