+++ src/main/java/com/solvd/app/services/PositionService.java
@@ -1,35 +1,30 @@
 package com.solvd.app.services;
 
-import com.solvd.app.daos.PositionDAO;
+import com.solvd.app.dao.PositionDAO;
 import com.solvd.app.models.Position;
 
-import java.sql.SQLException;
 import java.util.List;
 
 public class PositionService {
-    private PositionDAO positionDAO;
+    private PositionDAO positionDAO = new PositionDAO();
 
-    public PositionService(PositionDAO positionDAO) {
-        this.positionDAO = positionDAO;
-    }
-
-    public void createPosition(Position position) throws SQLException {
+    public void createPosition(Position position) {
         positionDAO.createEntity(position);
     }
 
-    public Position getPositionByID(int id) throws SQLException {
+    public Position getPositionByID(int id) {
         return positionDAO.getEntityByID(id);
     }
 
-    public void updatePosition(Position position) throws SQLException {
+    public void updatePosition(Position position) {
         positionDAO.updateEntity(position);
     }
 
-    public void deletePositionByID(int id) throws SQLException {
+    public void deletePositionByID(int id) {
         positionDAO.deleteEntityByID(id);
     }
 
-    public List<Position> getAllPositions() throws SQLException {
+    public List<Position> getAllPositions() {
         return positionDAO.getAll();
     }
 }
