+++ src/main/java/com/solvd/app/services/PrescriptionItemService.java
@@ -1,40 +1,35 @@
 package com.solvd.app.services;
 
-import com.solvd.app.daos.PrescriptionItemDAO;
+import com.solvd.app.dao.PrescriptionItemDAO;
 import com.solvd.app.models.Prescription;
 import com.solvd.app.models.PrescriptionItem;
 
-import java.sql.SQLException;
 import java.util.List;
 
 public class PrescriptionItemService {
-    private PrescriptionItemDAO prescriptionItemDAO;
+    private PrescriptionItemDAO prescriptionItemDAO = new PrescriptionItemDAO();
 
-    public PrescriptionItemService(PrescriptionItemDAO prescriptionItemDAO) {
-        this.prescriptionItemDAO = prescriptionItemDAO;
-    }
-
-    public void createPrescriptionItem(PrescriptionItem prescriptionItem) throws SQLException {
+    public void createPrescriptionItem(PrescriptionItem prescriptionItem) {
         prescriptionItemDAO.createEntity(prescriptionItem);
     }
 
-    public PrescriptionItem getPrescriptionItemByID(int id) throws SQLException {
+    public PrescriptionItem getPrescriptionItemByID(int id) {
         return prescriptionItemDAO.getEntityByID(id);
     }
 
-    public void updatePrescriptionItem(PrescriptionItem prescriptionItem) throws SQLException {
+    public void updatePrescriptionItem(PrescriptionItem prescriptionItem) {
         prescriptionItemDAO.updateEntity(prescriptionItem);
     }
 
-    public void deletePrescriptionItemByID(int id) throws SQLException {
+    public void deletePrescriptionItemByID(int id) {
         prescriptionItemDAO.deleteEntityByID(id);
     }
 
-    public List<PrescriptionItem> getAllPrescriptionItems() throws SQLException {
+    public List<PrescriptionItem> getAllPrescriptionItems() {
         return prescriptionItemDAO.getAll();
     }
 
-    public List<PrescriptionItem> getItemsByPrescription(Prescription prescription) throws SQLException {
+    public List<PrescriptionItem> getItemsByPrescription(Prescription prescription) {
         return prescriptionItemDAO.getItemsByPrescription(prescription);
     }
 }
