+++ src/main/java/com/solvd/app/services/StaffService.java
@@ -1,40 +1,35 @@
 package com.solvd.app.services;
 
-import com.solvd.app.daos.StaffDAO;
+import com.solvd.app.dao.StaffDAO;
 import com.solvd.app.models.Pharmacy;
 import com.solvd.app.models.Staff;
 
-import java.sql.SQLException;
 import java.util.List;
 
 public class StaffService {
-    private StaffDAO staffDAO;
+    private StaffDAO staffDAO = new StaffDAO();
 
-    public StaffService(StaffDAO staffDAO) {
-        this.staffDAO = staffDAO;
-    }
-
-    public void createStaff(Staff staff) throws SQLException {
+    public void createStaff(Staff staff) {
         staffDAO.createEntity(staff);
     }
 
-    public Staff getStaffByID(int id) throws SQLException {
+    public Staff getStaffByID(int id) {
         return staffDAO.getEntityByID(id);
     }
 
-    public void updateStaff(Staff staff) throws SQLException {
+    public void updateStaff(Staff staff) {
         staffDAO.updateEntity(staff);
     }
 
-    public void deleteStaffByID(int id) throws SQLException {
+    public void deleteStaffByID(int id) {
         staffDAO.deleteEntityByID(id);
     }
 
-    public List<Staff> getAllStaff() throws SQLException {
+    public List<Staff> getAllStaff() {
         return staffDAO.getAll();
     }
 
-    public List<Staff> getStaffByPharmacy(Pharmacy pharmacy) throws SQLException {
+    public List<Staff> getStaffByPharmacy(Pharmacy pharmacy) {
         return staffDAO.getStaffByPharmacy(pharmacy);
     }
 }
