+++ src/main/java/com/solvd/app/services/SupplierService.java
@@ -1,35 +1,30 @@
 package com.solvd.app.services;
 
-import com.solvd.app.daos.SupplierDAO;
+import com.solvd.app.dao.SupplierDAO;
 import com.solvd.app.models.Supplier;
 
-import java.sql.SQLException;
 import java.util.List;
 
 public class SupplierService {
-    private SupplierDAO supplierDAO;
+    private SupplierDAO supplierDAO = new SupplierDAO();
 
-    public SupplierService(SupplierDAO supplierDAO) {
-        this.supplierDAO = supplierDAO;
-    }
-
-    public void createSupplier(Supplier supplier) throws SQLException {
+    public void createSupplier(Supplier supplier) {
         supplierDAO.createEntity(supplier);
     }
 
-    public Supplier getSupplierByID(int id) throws SQLException {
+    public Supplier getSupplierByID(int id) {
         return supplierDAO.getEntityByID(id);
     }
 
-    public void updateSupplier(Supplier supplier) throws SQLException {
+    public void updateSupplier(Supplier supplier) {
         supplierDAO.updateEntity(supplier);
     }
 
-    public void deleteSupplierByID(int id) throws SQLException {
+    public void deleteSupplierByID(int id) {
         supplierDAO.deleteEntityByID(id);
     }
 
-    public List<Supplier> getAllSuppliers() throws SQLException {
+    public List<Supplier> getAllSuppliers() {
         return supplierDAO.getAll();
     }
 }
