+++ src/main/java/com/solvd/app/utils/ConnectionPool.java
@@ -1,44 +1,39 @@
 package com.solvd.app.utils;
 
 import java.sql.Connection;
-import java.util.ArrayList;
-import java.util.List;
+import java.util.Vector;
 
 public class ConnectionPool {
-private static ConnectionPool instance = null;
+    private static ConnectionPool instance = null;
     private static final int INITIAL_POOL_SIZE = 5;
     private static Vector<Connection> freeConnections = new Vector<>();
     private static Vector<Connection> usedConnections = new Vector<>();
 
-public static synchronized ConnectionPool getInstance() {
+    public static synchronized ConnectionPool getInstance() {
         if (instance == null) {
             instance = new ConnectionPool();
             create();
         }
         return instance;
     }
-    public ConnectionPool(int poolSize) {
-        this.connectionPool = new ArrayList<>(poolSize);
 
-        for (int i = 0; i < poolSize; i++) {
-            connectionPool.add(createNewConnectionForPool());
+    public static void create() {
+        for (int i = 0; i < INITIAL_POOL_SIZE; i++) {
+            freeConnections.add(createNewConnectionForPool());
         }
     }
 
-    private Connection createNewConnectionForPool() {
+    private static Connection createNewConnectionForPool() {
         return DBConnection.getConnection();
     }
 
     public synchronized Connection getConnection() {
-        int lastIndex = connectionPool.size() - 1;
-        if (lastIndex < 0) {
-            return createNewConnectionForPool();
-        } else {
-            return connectionPool.remove(lastIndex);
-        }
+        Connection connection = freeConnections.remove(freeConnections.size() - 1);
+        usedConnections.add(connection);
+        return connection;
     }
 
     public synchronized void releaseConnection(Connection connection) {
-        connectionPool.add(connection);
+        freeConnections.add(connection);
     }
 }
